Separate NoteSelectionBlueprint into a note piece
