+++ osu.Game.Rulesets.Mania/Edit/Blueprints/NoteSelectionBlueprint.cs
@@ -1,10 +1,9 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using osu.Framework.Allocation;
-using osu.Game.Graphics;
+using osu.Framework.Graphics;
+using osu.Game.Rulesets.Mania.Edit.Blueprints.Components;
 using osu.Game.Rulesets.Mania.Objects.Drawables;
-using osu.Game.Rulesets.Mania.Objects.Drawables.Pieces;
 
 namespace osu.Game.Rulesets.Mania.Edit.Blueprints
 {
@@ -13,18 +12,7 @@ namespace osu.Game.Rulesets.Mania.Edit.Blueprints
         public NoteSelectionBlueprint(DrawableNote note)
             : base(note)
         {
-            Scale = note.Scale;
-
-            CornerRadius = 5;
-            Masking = true;
-
-            AddInternal(new NotePiece());
-        }
-
-        [BackgroundDependencyLoader]
-        private void load(OsuColour colours)
-        {
-            Colour = colours.Yellow;
+            AddInternal(new EditNotePiece { RelativeSizeAxes = Axes.X });
         }
 
         protected override void Update()
